Fix crafting bug caused by goal ordering
The sequence of
    goals: [Siege Elite 2, Rosa Elite 2]
when crafting Bipolar Nanoflake results in the expected requirement of
14 Optimized Device (6 for Rosa, +8 to craft 8 Bipolar Nanoflakes),
but the goal ordering
    goals: [Rosa Elite 2, Siege Elite 2]
produced an erroneous requirement of 8 total Optimized Device due to
clobbering of the original requirements amount.

We have to first populate all the item requirements and then loop over
the active crafting recipes rather than trying to do it in one pass.