Replace goal autocomplete with Select
This led to really poor UI on mobile because the keyboard ends up
cutting away a huge portion of the available selection space, leading to
a ton of UI jank from the input increasing in height and the menu moving
to avoid it.

Next item is to make the Select render a browser-native element on
mobile only.