Fix incorrect counts for chain-crafted materials
In situations where an item is required both directly by a goal *and* as
an ingredient of a crafted item, when *that* item is crafted, its
crafted counts were incorrect.

Simple example is Rosa's Elite 2 which requires both Bipolar Nanoflake
and Optimized Device. Setting both to "craft" resulted in a doubling of
the materials required for Optimized Device.