Add failing test for chained crafting completion
A large enough quantity of lowest-tier material should be able to craft
everything up through the highest tier material; for example, 4 orirock
concentrations can be made (eventually) from 1000 orirock. However the
site incorrectly reports that only the orirock and orirock cube
requirements are met, while the orirock cluster and concentration
requirements are not.