Revert "Deduct ingredients when incrementing crafted items"
This reverts commit fb68deb990805a4f64fa9aa81bc39797eb4dc5fa.

On second thought, I don't like that the + button has different behavior
that is not immediately explained/obvious to the end user.

The better solution is to add a separate increment button when Crafting
is enabled that *crafts* an additional item (and reduces counts of its
ingredients) rather than just incrementing its stock. (A
crafting-decrement button would be unnecessary.)